+++ Active/assets/shaders/fluid/surface/vertex.glsl
@@ -4,11 +4,11 @@ in vec4 aVert;
 in vec2 aTex;
 in vec3 aNorm;
 
-out mediump vec4 vColor;
-out mediump vec2 vTex;
-out mediump vec3 vNorm;
-out mediump vec4 vFragPos;
-out mediump vec4 vModelPos;
+out highp vec4 vColor;
+out highp vec2 vTex;
+out highp vec3 vNorm;
+out highp vec4 vFragPos;
+out highp vec4 vModelPos;
 
 
 uniform mat4 uModelInvTrans;
